+++ command/cli_test.go
@@ -64,6 +64,30 @@ func (tc *testCliConfig) SaveStartCommandToProjectConfig(c string) error {
 	return nil
 }
 
+func (tc *testCliConfig) SaveDockerCliXdebugIniFilePath(c string) error {
+	return nil
+}
+
+func (tc *testCliConfig) SaveDockerFpmXdebugIniFilePath(c string) error {
+	return nil
+}
+
+func (tc *testCliConfig) SaveXDebugConifgLocaton(c string) error {
+	return nil
+}
+
+func (tc *testCliConfig) GetXDebugCliIniPath() string {
+	return ""
+}
+
+func (tc *testCliConfig) GetXDebugFpmIniPath() string {
+	return ""
+}
+
+func (tc *testCliConfig) GetXDebugConifgLocaton() string {
+	return ""
+}
+
 type testCliDialog struct{}
 
 func (d *testCliDialog) SetMainContaner([]string) (int, string, error) {
@@ -84,6 +108,18 @@ func (d *testCliDialog) DockerService() (string, error) {
 
 func (d *testCliDialog) DockerProjectPath(c string) (string, error) {
 	return "", nil
+}
+
+func (d *testCliDialog) DockerCliXdebugIniFilePath(c string) (string, error) {
+	return "", nil
+}
+
+func (d *testCliDialog) DockerFpmXdebugIniFilePath(c string) (string, error) {
+	return "", nil
+}
+
+func (d *testCliDialog) XDebugConfigLocation() (int, string, error) {
+	return 0, "", nil
 }
 
 type testCli struct {
@@ -168,7 +204,7 @@ func TestCliHandleCase3(t *testing.T) {
 			"cli": []string{"-it"},
 		},
 		command: map[string]string{
-			"cli": "bash",
+			"cli": "sh",
 		},
 	}
 
@@ -185,5 +221,5 @@ func TestCliHandleCase3(t *testing.T) {
 	args, err := cliCommandHandle("cli", cfg, dlg, cli, cl, a)
 
 	assert.Nil(t, err)
-	assert.Equal(t, []string{"exec", "-it", "containerName", "bash"}, args)
+	assert.Equal(t, []string{"exec", "-it", "containerName", "sh"}, args)
 }