+++ config/config.go
@@ -125,6 +125,11 @@ func (c *Config) GetProjectName() string {
 	return c.projectConfig.GetName()
 }
 
+// GetShell gets shell command
+func (c *Config) GetShell() string {
+	return c.projectConfig.GetShell()
+}
+
 // ProjectConfgFileFound checks if current path has config file
 func (c *Config) ProjectConfgFileFound() bool {
 	return c.hasProjectFile
@@ -159,6 +164,12 @@ func (c *Config) AddProjectConfigFile() (err error) {
 
 // main container
 
+// SaveShellCommand saves linux shell command
+func (c *Config) SaveShellCommand(cmd string) (err error) {
+	c.projectConfig.Shell = cmd
+	return c.saveProjectFile()
+}
+
 // SaveContainerNameToProjectConfig saves container name into project file
 func (c *Config) SaveContainerNameToProjectConfig(cn string) (err error) {
 	c.projectConfig.MainContainer = cn
@@ -170,9 +181,52 @@ func (c *Config) GetProjectMainContainer() string {
 	return c.projectConfig.GetMainContainer()
 }
 
+// Docker project path
+
 // GetProjectDockerPath gets project main container
 func (c *Config) GetProjectDockerPath() string {
 	return c.projectConfig.GetDockerProjectPath()
+}
+
+// SaveDockerProjectPath saves path to project in container into project file
+func (c *Config) SaveDockerProjectPath(path string) (err error) {
+	c.projectConfig.DockerProjectPath = path
+	return c.saveProjectFile()
+}
+
+// Xdebug
+
+// GetXDebugConifgLocaton gets cli xdebug ini file path
+func (c *Config) GetXDebugConifgLocaton() string {
+	return c.projectConfig.GetXDebugConifgLocaton()
+}
+
+// GetXDebugCliIniPath gets cli xdebug ini file path
+func (c *Config) GetXDebugCliIniPath() string {
+	return c.projectConfig.GetXDebugCliIniPath()
+}
+
+// GetXDebugFpmIniPath gets fpm xdebug ini file path
+func (c *Config) GetXDebugFpmIniPath() string {
+	return c.projectConfig.GetXDebugFpmIniPath()
+}
+
+// SaveXDebugConifgLocaton saves xdebug file location
+func (c *Config) SaveXDebugConifgLocaton(path string) (err error) {
+	c.projectConfig.XDebugLocation = path
+	return c.saveProjectFile()
+}
+
+// SaveDockerCliXdebugIniFilePath saves xdebug cli ini file path into project file
+func (c *Config) SaveDockerCliXdebugIniFilePath(path string) (err error) {
+	c.projectConfig.XDebugCliIniPath = path
+	return c.saveProjectFile()
+}
+
+// SaveDockerFpmXdebugIniFilePath saves xdebug fpm ini file path into project file
+func (c *Config) SaveDockerFpmXdebugIniFilePath(path string) (err error) {
+	c.projectConfig.XDebugFpmIniPath = path
+	return c.saveProjectFile()
 }
 
 // main start command
@@ -188,13 +242,7 @@ func (c *Config) SaveStartCommandToProjectConfig(cmd string) (err error) {
 	return c.saveProjectFile()
 }
 
-// SaveDockerProjectPath saves path to project in container into project file
-func (c *Config) SaveDockerProjectPath(path string) (err error) {
-	c.projectConfig.DockerProjectPath = path
-	return c.saveProjectFile()
-}
-
-// GetFile fsfdsafsd
+// GetFile gets project file
 func (c *Config) GetFile() string {
 	return c.ProjectFile
 }
@@ -206,6 +254,8 @@ func (c *Config) saveProjectFile() error {
 func (c *Config) getProjectFile() string {
 	return strings.TrimRight(c.projectConfig.GetPath(), string(os.PathSeparator)) + string(os.PathSeparator) + c.ProjectFile
 }
+
+// Copyright
 
 // EnableCopyright Enable copyright output
 func (c *Config) EnableCopyright() error {
@@ -223,6 +273,8 @@ func (c *Config) DisableCopyright() error {
 func (c *Config) ShowCopyrightText() bool {
 	return c.globalConfig.ShowCopyrightText()
 }
+
+// Docker instance command
 
 // SetDockerCommand define docker command
 func (c *Config) SetDockerCommand(command string) error {